Fold the host-only sources back into their shared device siblings
barrier_classes.cpp and distance_type_host.cpp existed only to keep code
away from nvcc, which left barrier.hpp and distance_type.hpp each
implemented across two .cpp files. The dual-compilation mechanism does not
require a separate file, only that the host-only code be invisible to the
device pass, so both are merged back behind `#ifndef __CUDACC__`.

- barrier.cpp regains the virtual Barrier hierarchy. Virtual dispatch
  cannot cross the host/device boundary -- a host-built vtable holds host
  code addresses, CUDA forbids passing an object of a class with virtual
  functions to a __global__ function, and BarrierPotential owns its
  barrier through a host-only std::shared_ptr -- so the guard also keeps
  every class symbol, float and double included, in the host object.
- distance_type.cpp regains the error reporting helpers. Here the blocker
  is that nvcc cannot parse spdlog, so the guard covers the includes as
  well as the definitions; an #include in a false branch is never
  processed. Every use of the logger and of fmt must stay inside it.

The free functions in both files are unchanged and still emit relocatable
device code: cuobjdump confirms all six barrier functions in the fatbin,
and the device object holds zero class symbols and zero spdlog symbols.
