+++ src/ipc/distance/distance_type.cpp
@@ -6,6 +6,19 @@
 
 #include <limits>
 
+// The error reporting helpers at the bottom of this file are host only: nvcc
+// cannot parse spdlog, so their definitions and the headers they need both sit
+// behind `#ifndef __CUDACC__` and are compiled by the host pass alone. Keep
+// every use of the logger and of fmt inside those blocks -- an unguarded one
+// fails as a parse error deep inside fmt rather than as a missing declaration.
+#ifndef __CUDACC__
+#include <ipc/utils/logger.hpp>
+
+#include <spdlog/spdlog.h>
+
+#include <stdexcept>
+#endif
+
 namespace ipc::detail {
 
 template <typename T>
@@ -221,5 +234,37 @@ template EdgeEdgeDistanceType edge_edge_parallel_distance_type<float>(Eigen::Con
 template EdgeEdgeDistanceType edge_edge_parallel_distance_type<double>(Eigen::ConstRef<Eigen::Vector3d>, Eigen::ConstRef<Eigen::Vector3d>, Eigen::ConstRef<Eigen::Vector3d>, Eigen::ConstRef<Eigen::Vector3d>);
 #endif
 // clang-format on
+
+// ============================================================================
+// Error reporting -- host only.
+// ============================================================================
+//
+// The inline wrappers in distance_type.hpp call these on the host and trap on
+// the device, so the device pass needs neither the definitions nor spdlog.
+#ifndef __CUDACC__
+
+void warn_degenerate_point_edge_host() noexcept
+{
+    logger().warn("Degenerate edge in point_edge_distance_type!");
+}
+
+void throw_invalid_distance_type_host(const char* function)
+{
+    throw std::invalid_argument(
+        fmt::format("{}: invalid distance type", function));
+}
+
+void throw_auto_requires_explicit_dtype_host(const char* function)
+{
+    throw std::invalid_argument(
+        fmt::format(
+            "{}: an explicit distance type is required for non-floating-point "
+            "scalars; resolving AUTO means comparing single ordered values, "
+            "which an autodiff, SIMD batch, or interval scalar does not "
+            "provide",
+            function));
+}
+
+#endif // !__CUDACC__
 
 } // namespace ipc::detail
